Format list patterns (#57568)
