+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Formatting/Rules/IndentBlockFormattingRule.cs
@@ -50,6 +50,8 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
 
             AddBlockIndentationOperation(list, node);
 
+            AddBracketIndentationOperation(list, node);
+
             AddLabelIndentationOperation(list, node);
 
             AddSwitchIndentationOperation(list, node);
@@ -212,7 +214,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
             var bracePair = node.GetBracePair();
 
             // don't put block indentation operation if the block only contains label statement
-            if (!bracePair.IsValidBracePair())
+            if (!bracePair.IsValidBracketOrBracePair())
             {
                 return;
             }
@@ -242,6 +244,24 @@ namespace Microsoft.CodeAnalysis.CSharp.Formatting
             }
 
             AddIndentBlockOperation(list, bracePair.openBrace.GetNextToken(includeZeroWidth: true), bracePair.closeBrace.GetPreviousToken(includeZeroWidth: true));
+        }
+
+        private static void AddBracketIndentationOperation(List<IndentBlockOperation> list, SyntaxNode node)
+        {
+            var bracketPair = node.GetBracketPair();
+
+            if (!bracketPair.IsValidBracketOrBracePair())
+            {
+                return;
+            }
+
+            if (node.IsKind(SyntaxKind.ListPattern) && node.Parent != null)
+            {
+                // Brackets in list patterns are formatted like blocks, so align close bracket with open bracket
+                AddAlignmentBlockOperationRelativeToFirstTokenOnBaseTokenLine(list, bracketPair);
+
+                AddIndentBlockOperation(list, bracketPair.openBracket.GetNextToken(includeZeroWidth: true), bracketPair.closeBracket.GetPreviousToken(includeZeroWidth: true));
+            }
         }
 
         private static void AddAlignmentBlockOperationRelativeToFirstTokenOnBaseTokenLine(List<IndentBlockOperation> list, (SyntaxToken openBrace, SyntaxToken closeBrace) bracePair)
